+++ nucypher-core/src/reencryption.rs
@@ -126,18 +126,19 @@ fn signed_message(capsules: &[Capsule], cfrags: &[CapsuleFrag]) -> Vec<u8> {
 
 impl ReencryptionResponse {
     /// Creates and signs a new reencryption response.
-    pub fn new(
+    pub fn new<'a>(
         signer: &Signer,
-        capsules: &[Capsule],
-        vcfrags: impl IntoIterator<Item = VerifiedCapsuleFrag>,
+        capsules_and_vcfrags: impl IntoIterator<Item = (&'a Capsule, VerifiedCapsuleFrag)>,
     ) -> Self {
+        let (capsules, vcfrags): (Vec<_>, Vec<_>) = capsules_and_vcfrags.into_iter().unzip();
+
         // un-verify
         let cfrags: Vec<_> = vcfrags
             .into_iter()
             .map(|vcfrag| vcfrag.unverify())
             .collect();
 
-        let signature = signer.sign(&signed_message(capsules, &cfrags));
+        let signature = signer.sign(&signed_message(&capsules, &cfrags));
 
         ReencryptionResponse {
             cfrags: cfrags.into_boxed_slice(),